feat(#77): MyStudyRepository, MyStudySettingRepository 정의
- MyStudyRepository 2개로 쪼개기
